refactor(core): flatten core/__tests__ — rename 5 parallel suites (#2565 slice 7)
core/__tests__ had 5 files with basename collisions, each covering a
different angle than its top-level counterpart:

- artifact.test.ts (top:372 / nested:511) — both test Artifact, nested
  has more cases. Renamed nested to artifact-extended.test.ts.
- logger.test.ts (top:252 / nested:353) — top covers sanitize/
  sanitizeDeep/createLogger/setGlobalLogLevel; nested is dedicated to
  credential-pattern detection (API keys, Bearer, AWS/Azure/GCP, GitHub
  tokens). Renamed nested to logger-credentials.test.ts.
- result.test.ts (top:223 / nested:141) — top covers ok/err/isOk/isErr/
  map; nested wraps in a single Result describe and adds unwrap/unwrapOr
  FP helpers. Renamed nested to result-fp.test.ts.
- trace-exporter.test.ts (top:178 / nested:309) — top covers 3 export
  functions; nested adds exportToFile, printTrace, generateTraceFilename.
  Renamed nested to trace-exporter-formats.test.ts.
- trace.test.ts (top:1233 / nested:617) — both test Tracer class; top
  is much larger. Renamed nested to trace-extended.test.ts.

Fix ../X static + ../../X dynamic imports → ./X (or ../X for cross-
package). Remove the empty __tests__/ directory.

All 1171 core tests still pass.
